Made timeouts in Dispatchables UInts not UInt32s

+++ DynamicTimeSlotDispatchableItem.swift
@@ -59,14 +59,12 @@
 public class DynamicTimeSlotDispatchableItem: Dispatchable {
 
     public let item: RunnableMachine
-    public var timeout: UInt32 {
+    public var timeout: UInt {
         let wcet: UInt = self.item.worstCaseExecutionTime
         if (0 == wcet) {
-            return UInt32(self.timeslot)
+            return self.timeslot
         }
-        return UInt32(
-            (self.timeslot / item.worstCaseExecutionTime + 1) * self.timeslot
-        )
+        return (self.timeslot / item.worstCaseExecutionTime + 1) * self.timeslot
     }
     private let timeslot: UInt
